requirements.txt
Django==5.0.7
djangorestframework==3.15.2
dj-database-url==2.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0
